# investment_mcp_api/docs/openapi/investment_mcp_api.yaml
openapi: 3.0.3
info:
  title: Investment MCP API
  description: |
    Swedish Investment Analysis API with Historical Data and AI-Powered Recommendations
    
    This API provides comprehensive investment analysis for Swedish markets, including:
    - 20+ years of historical fund data analysis
    - Real-time Swedish economic indicators
    - AI-powered portfolio optimization
    - Risk assessment and stress testing
    - Market regime analysis
    
    ## Authentication
    API keys are required for all endpoints. Contact your administrator for access.
    
    ## Rate Limiting
    - Standard tier: 100 requests per minute
    - Premium tier: 1000 requests per minute
    
    ## Data Sources
    - Riksbank (Swedish Central Bank) for economic data
    - SCB (Statistics Sweden) for demographic and economic statistics  
    - Yahoo Finance and other providers for historical fund data
    
  version: 3.0.0
  contact:
    name: Investment MCP API Support
    email: support@investment-mcp.com
    url: https://docs.investment-mcp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.investment-mcp.com/v1
    description: Production server
  - url: https://staging-api.investment-mcp.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /portfolio/analysis:
    post:
      tags:
        - Portfolio Analysis
      summary: Generate AI-powered portfolio analysis
      description: |
        Generate comprehensive portfolio analysis using historical data and Swedish economic context.
        Includes optimal allocation recommendations, risk assessment, and stress testing.
      operationId: analyzePortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioAnalysisRequest'
            examples:
              conservative:
                summary: Conservative investor analysis
                value:
                  risk_profile: "conservative"
                  investment_amount: 500000
                  investment_horizon_years: 10
                  current_allocation: {}
              aggressive:
                summary: Aggressive investor analysis  
                value:
                  risk_profile: "aggressive"
                  investment_amount: 1000000
                  investment_horizon_years: 15
                  current_allocation:
                    "AVANZA_USA": 0.3
                    "DNB_GLOBAL_INDEKS_S": 0.2
      responses:
        '200':
          description: Portfolio analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioAnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /portfolio/allocations/historical:
    post:
      tags:
        - Portfolio Analysis
      summary: Generate historical allocation analysis
      description: Generate optimal allocation based on 20+ years of historical data analysis
      operationId: generateHistoricalAllocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoricalAllocationRequest'
      responses:
        '200':
          description: Historical allocation analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalAllocationResponse'
  /portfolio/stress-test:
    post:
      tags:
        - Portfolio Analysis
      summary: Perform portfolio stress testing
      description: Test portfolio performance against historical crisis scenarios
      operationId: stressTestPortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StressTestRequest'
      responses:
        '200':
          description: Stress test completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StressTestResponse'

  # Fund Data Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    PortfolioAnalysisRequest:
      type: object
      required:
        - risk_profile
        - investment_amount
      properties:
        risk_profile:
          type: string
          enum: [conservative, balanced, aggressive]
          description: Investor risk tolerance level
        investment_amount:
          type: number
          minimum: 10000
          maximum: 100000000
          description: Investment amount in SEK
        investment_horizon_years:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: Investment time horizon in years
        current_allocation:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
          description: Current portfolio allocation (fund_code -> weight)
        constraints:
          type: object
          properties:
            max_funds:
              type: integer
              minimum: 1
              maximum: 12
              default: 8
            min_allocation_per_fund:
              type: number
              minimum: 0.01
              maximum: 0.5
              default: 0.05
            exclude_funds:
              type: array
              items:
                type: string
    HistoricalAllocationRequest:
      type: object
      required:
        - risk_profile
        - investment_horizon_years
      properties:
        risk_profile:
          type: string
          enum: [conservative, balanced, aggressive]
        investment_horizon_years:
          type: integer
          minimum: 1
          maximum: 50
        include_stress_test:
          type: boolean
          default: true
        historical_years:
          type: integer
          minimum: 5
          maximum: 25
          default: 20
    StressTestRequest:
      type: object
      required:
        - allocation
      properties:
        allocation:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
          description: Portfolio allocation to test
        scenarios:
          type: array
          items:
            type: string
            enum: [2008_crisis, covid_2020, dotcom_2000, all]
          default: ["all"]
...
    PortfolioAnalysisResponse:
      type: object
      properties:
        allocation:
          type: object
          additionalProperties:
            type: number
          description: Recommended allocation (fund_code -> weight)
        expected_metrics:
          $ref: '#/components/schemas/PortfolioMetrics'
        historical_analysis:
          $ref: '#/components/schemas/HistoricalAnalysis'
        stress_test:
          $ref: '#/components/schemas/StressTestResults'
        swedish_economic_context:
          $ref: '#/components/schemas/SwedishEconomicContext'
        ai_reasoning:
          type: object
          properties:
            allocation_rationale:
              type: string
            swedish_economic_rationale:
              type: string
            risk_assessment:
              type: string
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
        generated_at:
          type: string
          format: date-time
    HistoricalAllocationResponse:
      type: object
      properties:
        allocation:
          type: object
          additionalProperties:
            type: number
        historical_reasoning:
          type: string
        performance_analysis:
          $ref: '#/components/schemas/HistoricalPerformance'
        regime_analysis:
          $ref: '#/components/schemas/RegimeAnalysis'
    StressTestResponse:
      type: object
      properties:
        scenarios:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/StressTestScenario'
        overall_risk_score:
          type: number
          minimum: 0
          maximum: 100
        recommendations:
          type: array
          items:
            type: string
...
    PortfolioMetrics:
      type: object
      properties:
        expected_annual_return:
          type: number
        expected_volatility:
          type: number
        expected_sharpe_ratio:
          type: number
        expected_max_drawdown:
          type: number
        diversification_ratio:
          type: number
    HistoricalAnalysis:
      type: object
      properties:
        performance_periods:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/PeriodPerformance'
        correlation_analysis:
          type: object
        regime_performance:
          $ref: '#/components/schemas/RegimeAnalysis'
    HistoricalPerformance:
      type: object
      properties:
        total_return:
          type: object
          additionalProperties:
            type: number
        annualized_return:
          type: number
        volatility:
          type: number
        best_year:
          type: number
        worst_year:
          type: number
        positive_years_percentage:
          type: number
    RegimeAnalysis:
      type: object
      properties:
        bull_market:
          $ref: '#/components/schemas/RegimePerformance'
        bear_market:
          $ref: '#/components/schemas/RegimePerformance'
        high_volatility:
          $ref: '#/components/schemas/RegimePerformance'
        low_volatility:
          $ref: '#/components/schemas/RegimePerformance'
    RegimePerformance:
      type: object
      properties:
        average_return:
          type: number
        volatility:
          type: number
        frequency:
          type: number
        max_drawdown:
          type: number
    StressTestResults:
      type: object
      properties:
        scenarios:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/StressTestScenario'
        worst_case_loss:
          type: number
        recovery_time_estimate:
          type: string
    StressTestScenario:
      type: object
      properties:
        scenario_name:
          type: string
        portfolio_return:
          type: number
        duration_days:
          type: integer
        description:
          type: string
    SwedishEconomicContext:
      type: object
      properties:
        current_phase:
          type: string
        key_rates:
          type: object
          properties:
            repo_rate:
              type: number
            inflation:
              type: number
        investment_implications:
          type: array
          items:
            type: string
...
    PeriodPerformance:
      type: object
      properties:
        total_return:
          type: number
        annualized_return:
          type: number
        volatility:
          type: number
        sharpe_ratio:
          type: number
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid API key"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Rate limit exceeded"
              retry_after:
                type: integer
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              request_id:
                type: string
...